Compare production to QC versions of an SDTM table

Wraps compareDF::compare_df() function and prints a high level summary using summary_diffs(). The arguments of compareDF::compare_df() are set to stop_on_error = F to avoid errors if the data frames are the same, and group_col is set to key_vars.

Usage

compare_qc_to_prod(qc, prod, group_col)

Arguments

qc

a data frame, the SDTM QC table

prod

a data frame, the SDTM production table

group_col

a character vector, the columns to use a unique keys for comparing qc and prod. Ideally this is the key variables for the domain as set out in the specification.

Value

a named list as returned by compareDF::compare_df(). A side effect also prints a high level summary using summary_diffs().